pyproject.toml
[project]
name = "lacrei-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "capf01",email = "cesaraugustopachecoferreira@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
